pyproject.toml
[project]
name = "qa-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.15.3",
    "factory-boy>=3.3.3",
    "faker>=40.15.0",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.3",
    "pytest-dotenv>=0.5.2",
    "pytest-httpserver>=1.1.5",
    "pytest-playwright>=0.7.2",
    "pytest-rerunfailures>=16.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.11",
    "sqlalchemy>=2.0.49",
    "testcontainers>=4.14.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
reruns = 2
reruns_delay = 1
addopts = "--alluredir=reports/allure-results"
markers = [
    "ui: marks tests as UI tests",
    "api: marks tests as API tests",
    "smoke: marks tests as smoke tests",
    "regression: marks tests as regression tests",
]

[tool.ruff]
line-length = 150
